# xml-doc-flow-backend/docs/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - basicAuth: []
tags:
  - name: 01. API загрузки строительных документов
    description: Загрузка XML, валидация по XSD, версионирование и выдача документов
paths:
  /api/documents/{id}/replace:
    put:
      tags:
        - 01. API загрузки строительных документов
      summary: 'Замена XML: новая версия при совпадении типа и номера документа'
      operationId: replace
      parameters:
        - name: id
          in: path
          description: Id существующей версии документа, от которой ведётся замена
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  description: Новый XML той же схемы с тем же номером документа
                  format: binary
      responses:
        '409':
          description: 'Конфликт (validationStatus=INVALID_CONFLICT): другой тип/номер относительно выбранной версии'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadDocumentResponse'
        '404':
          description: Версия с указанным id не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: 'Новая версия сохранена. При `validationStatus=VALID` — XSD пройден; при `INVALID_*` — ошибки в `validationErrors`. HTTP 400 не используется для результата валидации'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadDocumentResponse'
        '400':
          description: Ошибка чтения файла из multipart (`ErrorResponse`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/documents:
    post:
      tags:
        - 01. API загрузки строительных документов
      summary: 'Загрузка XML: определение типа по корню, XSD-валидация, сохранение'
      operationId: upload
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  description: Файл экземпляра XML-документа (не XSL/XSD/SVG)
                  format: binary
      responses:
        '400':
          description: Не удалось прочитать файл из multipart (`MultipartFileReadException`, тело `ErrorResponse`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: 'Версия документа сохранена в БД (BYTEA). При `validationStatus=VALID` — XSD пройден; при `INVALID_*` — в теле `validationErrors` и `valid=false`. HTTP 400 не используется для результата валидации'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadDocumentResponse'
  /api/documents/search:
    post:
      tags:
        - 01. API загрузки строительных документов
      summary: Поиск документов с фильтрацией и пагинацией (JSON)
      operationId: search
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentSearchRequest'
        required: true
      responses:
        '400':
          description: Невалидные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: Страница документов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPageResponse'
  /api/documents/{id}:
    get:
      tags:
        - 01. API загрузки строительных документов
      summary: 'Карточка логического документа: объект, участники, версии, кто загрузил'
      operationId: getById
      parameters:
        - name: id
          in: path
          description: Идентификатор логического документа (documents.id), не id версии
          required: true
          schema:
            type: integer
            format: int64
          example: 13
      responses:
        '404':
          description: Документ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: Полная карточка по идентификатору логического документа (documents.id)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentCardResponse'
  /api/documents/{id}/xml:
    get:
      tags:
        - 01. API загрузки строительных документов
      summary: Скачивание исходного XML по id версии
      operationId: downloadXml
      parameters:
        - name: id
          in: path
          description: Идентификатор версии документа для скачивания XML
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '404':
          description: Документ не найден
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: 'Тело ответа — содержимое XML, Content-Disposition: attachment'
          content:
            application/xml: {}
  /api/documents/{documentId}/versions:
    get:
      tags:
        - 01. API загрузки строительных документов
      summary: Список версий логического документа
      operationId: listVersions
      parameters:
        - name: documentId
          in: path
          description: Идентификатор логического документа (documents.id)
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '404':
          description: Документ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: Версии в порядке убывания номера версии
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VersionInfo'
  /api/documents/{documentId}/versions/{versionId}/download:
    get:
      tags:
        - 01. API загрузки строительных документов
      summary: Скачивание XML конкретной версии документа
      operationId: downloadXmlByDocumentAndVersion
      parameters:
        - name: documentId
          in: path
          description: Идентификатор логического документа (documents.id)
          required: true
          schema:
            type: integer
            format: int64
        - name: versionId
          in: path
          description: Идентификатор версии (document_versions.id)
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Тело ответа — содержимое XML
          content:
            application/xml: {}
        '404':
          description: Документ или версия не найдены, либо версия не относится к документу
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/documents/{documentId}/versions/latest/download:
    get:
      tags:
        - 01. API загрузки строительных документов
      summary: Скачивание XML последней версии документа
      operationId: downloadLatestXml
      parameters:
        - name: documentId
          in: path
          description: Идентификатор логического документа (documents.id)
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Тело ответа — содержимое XML
          content:
            application/xml: {}
        '404':
          description: Документ или версии не найдены
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    UploadDocumentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        documentId:
          type: integer
          format: int64
        docType:
          type: string
        documentNumber:
          type: string
        version:
          type: integer
          format: int32
        uploadedAt:
          type: string
          format: date-time
        valid:
          type: boolean
        validationStatus:
          type: string
          enum:
            - VALID
            - INVALID_SCHEMA
            - INVALID_UNKNOWN_DOCUMENT_TYPE
            - INVALID_DOCUMENT_ROOT
            - INVALID_XML
            - INVALID_CONFLICT
        validationErrors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDto'
    ValidationErrorDto:
      type: object
      properties:
        message:
          type: string
        lineNumber:
          type: integer
          format: int32
        columnNumber:
          type: integer
          format: int32
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
    DocumentSearchRequest:
      type: object
      properties:
        objectId:
          type: integer
          format: int64
        objectCode:
          maxLength: 128
          minLength: 0
          type: string
        documentType:
          maxLength: 1024
          minLength: 0
          type: string
        documentNumber:
          maxLength: 512
          minLength: 0
          type: string
        status:
          maxLength: 32
          minLength: 0
          type: string
        validationStatus:
          maxLength: 32
          minLength: 0
          type: string
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
        uploadedFrom:
          type: string
          format: date-time
        uploadedTo:
          type: string
          format: date-time
        latestOnly:
          type: boolean
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          maximum: 500
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
        statusOrDefault:
          type: string
        latestOnlyOrDefault:
          type: boolean
    ConstructionObjectDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        objectCode:
          type: string
        objectName:
          type: string
        address:
          type: string
    DocumentLinks:
      type: object
      properties:
        downloadUrl:
          type: string
        versionsUrl:
          type: string
        selfUrl:
          type: string
    DocumentPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
        first:
          type: boolean
    DocumentResponse:
      type: object
      properties:
        versionId:
          type: integer
          format: int64
        documentId:
          type: integer
          format: int64
        documentNumber:
          type: string
        documentDate:
          type: string
          format: date
        title:
          type: string
        documentType:
          $ref: '#/components/schemas/DocumentTypeDto'
        constructionObject:
          $ref: '#/components/schemas/ConstructionObjectDto'
        currentVersion:
          type: integer
          format: int32
        latestVersion:
          type: boolean
        validationStatus:
          type: string
        status:
          type: string
        uploadedBy:
          $ref: '#/components/schemas/UserDto'
        uploadedAt:
          type: string
          format: date-time
        xmlFileName:
          type: string
        xmlFileSize:
          type: integer
          format: int64
        links:
          $ref: '#/components/schemas/DocumentLinks'
    DocumentTypeDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        typeCode:
          type: string
        typeName:
          type: string
        category:
          type: string
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        fullName:
          type: string
    ConstructionObjectAddressDto:
      type: object
      properties:
        country:
          type: string
        region:
          type: string
        district:
          type: string
        locality:
          type: string
        street:
          type: string
        house:
          type: string
        postalCode:
          type: string
        oktmo:
          type: string
        fullAddress:
          type: string
    ConstructionObjectInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          $ref: '#/components/schemas/ConstructionObjectAddressDto'
        objectCode:
          type: string
    DocumentCardResponse:
      type: object
      properties:
        documentId:
          type: integer
          format: int64
        documentNumber:
          type: string
        documentDate:
          type: string
          format: date
        docType:
          type: string
        docTypeName:
          type: string
        constructionObject:
          $ref: '#/components/schemas/ConstructionObjectInfo'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantInfo'
        currentVersion:
          type: integer
          format: int32
        uploadedAt:
          type: string
          format: date-time
        uploadedBy:
          $ref: '#/components/schemas/UserInfo'
        versions:
          type: array
          items:
            $ref: '#/components/schemas/VersionInfo'
        status:
          type: string
        validationStatus:
          type: string
    OrganizationInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        inn:
          type: string
        ogrn:
          type: string
        address:
          $ref: '#/components/schemas/ConstructionObjectAddressDto'
    ParticipantInfo:
      type: object
      properties:
        role:
          type: string
        roleName:
          type: string
        organization:
          $ref: '#/components/schemas/OrganizationInfo'
        address:
          $ref: '#/components/schemas/ConstructionObjectAddressDto'
    UserInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        fullName:
          type: string
    VersionInfo:
      type: object
      properties:
        versionId:
          type: integer
          format: int64
        versionNumber:
          type: integer
          format: int32
        uploadedAt:
          type: string
          format: date-time
        validationStatus:
          type: string
        isCurrent:
          type: boolean
        downloadUrl:
          type: string
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic