# xml-doc-flow-backend/services/app/src/main/java/ru/artwell/contractor/dto/UserInfo.java
package ru.artwell.contractor.dto;

import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
public class UserInfo {
    private Long id;
    private String username;
    private String fullName;

    public UserInfo(Long id, String username, String fullName) {
        this.id = id;
        this.username = username;
        this.fullName = fullName;
    }
}